+++ docs/changes.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+2.0.0 (2019/XX)
+---------------
+
+Improvements:
+
+- ``Apyfal.aws.AWSHost`` now support and use of spot instances by default.
+  This can be disabled with the ``spot_instance=False`` parameter.
+  Spot block duration can be specified using the ``spot_block_duration``
+  parameter.
+
 1.2.7 (2019/04)
 ---------------